@@ -9,13 +9,11 @@
     "express-graphql": "^0.7.1",
     "graphql": "^14.0.2",
     "knex": "^0.16.3",
-    "lodash": "^4.17.11",
+    "lodash": "^4.17.15",
     "pg": "^7.7.1",
     "prop-types": "^15.6.2",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-draggable": "^3.1.1",
-    "react-helmet": "^5.2.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-redux": "^6.0.0",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
