@@ -1,37 +1,32 @@
 import { useState } from 'react';
 import { jsx } from '@emotion/core'; /** @jsx jsx */
-import { OLL } from '.';
-import { Link, FlexContainer, FlexItem, Image } from '../particles';
-// import SidebarToggle from './SidebarToggle';
+import OLL from './OLL';
+import { FlexContainer, FlexItem, Image } from '../particles';
 
-// import { topNavCss
-
-const Algorithm = ({ algorithm = {}, step }) => {
-  return (
-    <article>
-      <FlexContainer>
-        <FlexItem>
-          <h2>{step} {algorithm.id}</h2>
-          <Image
-            src={algorithm.img}
-            width={80}
-            height={80}
-            external
-          />
-        </FlexItem>
-        <FlexItem>
-          <h2>Name: {algorithm.name}</h2>
-          {algorithm.algos.map((a, index) => (
-            <h3 key={index}>{a}</h3>
-          ))}
-        </FlexItem>
-        <FlexItem>
-          Squares: {algorithm.squares}
-        </FlexItem>
-      </FlexContainer>
-    </article>
-  );
-};
+const Algorithm = ({ algorithm = {}, step }) => (
+  <article>
+    <FlexContainer>
+      <FlexItem>
+        <h2>{step} {algorithm.id}</h2>
+        <Image
+          src={algorithm.img}
+          width={80}
+          height={80}
+          external
+        />
+      </FlexItem>
+      <FlexItem>
+        <h2>Name: {algorithm.name}</h2>
+        {algorithm.algos.map((a, index) => (
+          <h3 key={index}>{a}</h3>
+        ))}
+      </FlexItem>
+      <FlexItem>
+        Squares: {algorithm.squares}
+      </FlexItem>
+    </FlexContainer>
+  </article>
+);
 
 const Algorithms = ({ algorithms = [], step }) => (
   <section>
@@ -42,8 +37,8 @@ const Algorithms = ({ algorithms = [], step }) => (
 );
 
 const Cube = () => {
-  // const [step, filterStep] = useState('OLL');
-  const step = 'OLL';
+  const [step, filterStep] = useState('OLL');
+
   return (
     <section>
       <Algorithms algorithms={OLL} step={step} />
