@@ -1,19 +1,10 @@
-import React from 'react';
-import { jsx } from '@emotion/core'; /** @jsx jsx */
+// import { jsx } from '@emotion/core'; /** @jsx jsx */
 // import { xiangqiCss } from './xiangqiCss';
-import Board from './Board';
+// import Board from './Board';
 
-class Xiangqi extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-      //
-    };
-  }
-
-  render() {
-    return <Board />;
-  }
-}
+// const Xiangqi = () => <Board />;
+// temporarily exporting nothing so I can remove react-draggable.
+// when using Board, re-install react-draggable.
+const Xiangqi = () => {};
 
 export default Xiangqi;
